Modified the example to work with JPA 2 and JTA 1.2
RecipeBookServiceDAOImpl modified to use annotations.
persistence.xml modified to remove the non-jta-datasource reference and the openjpa connection url refers to 'pi' instead of localhost.
blueprint.xml modified to remove the service declarations of ClientDataSource and ClientXADataSource. These are now created by using the pax-jdbc family of bundles and creating a configuration file to deploy to Karaf,